+++ example/src/app/_layout.tsx
@@ -1,15 +1,22 @@
 import '../global.css'
-import { SafeAreaProvider } from 'react-native-safe-area-context';
+import { SafeAreaListener, SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
 import { Stack } from 'expo-router';
 import { View } from 'react-native';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
+import { Uniwind } from 'uniwind';
+import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
 
 const RootLayout = () => {
   return (
-    <SafeAreaProvider>
-      <View className="flex-1">
-        <Stack screenOptions={{ headerLargeTitleEnabled: true, headerBackButtonDisplayMode: 'minimal' }} />
-      </View>
-    </SafeAreaProvider>
+    <GestureHandlerRootView>
+      <BottomSheetModalProvider>
+        <SafeAreaListener onChange={({ insets }) => Uniwind.updateInsets(insets)}>
+          <View className="flex-1">
+            <Stack screenOptions={{ headerLargeTitleEnabled: true, headerBackButtonDisplayMode: 'minimal' }} />
+          </View>
+        </SafeAreaListener>
+      </BottomSheetModalProvider>
+    </GestureHandlerRootView>
   );
 }